--- templates/index.html
@@ -1,14 +1,0 @@
-<!-- Key:
-    - image_path
-    - options 
--->
-
-<!DOCTYPE html>
-<html>
-<body>
-    <img src="{{ image_path }}" alt="image">
-    {% for option in options %}
-        <p>{{ option }}</p>
-    {% endfor %}
-</body>
-</html>
